fix(friendRec): event recommendations and events module cleanup
- Wire getEvents to load per-user recommendations from Convex
- Add events/utils (serialization, H3 index, rec scoring)
- Trim duplicate logic in events/queries; fix ingest H3 import
- Map search overlay + theme-toggle lint fixes
