+++ packages/backend/convex/_generated/api.d.ts
@@ -22,6 +22,7 @@ import type * as events_ingest from "../events/ingest.js";
 import type * as events_mutations from "../events/mutations.js";
 import type * as events_queries from "../events/queries.js";
 import type * as events_search from "../events/search.js";
+import type * as events_utils from "../events/utils.js";
 import type * as files from "../files.js";
 import type * as friends from "../friends.js";
 import type * as http from "../http.js";
@@ -59,6 +60,7 @@ declare const fullApi: ApiFromModules<{
   "events/mutations": typeof events_mutations;
   "events/queries": typeof events_queries;
   "events/search": typeof events_search;
+  "events/utils": typeof events_utils;
   files: typeof files;
   friends: typeof friends;
   http: typeof http;
