+++ packages/backend/convex/events/queries.ts
@@ -1,53 +1,19 @@
-import { latLngToCell } from 'h3-js';
-
 import { v } from 'convex/values';
 import type { Doc } from '../_generated/dataModel';
 import { query, type QueryCtx } from '../_generated/server';
 import { __backend_only_guestOrAuthenticatedUser } from '../auth';
-import { getThreadedCommentsByPost } from '../comments';
 import { getHiddenUserIds } from '../moderation/block';
 import { getHiddenPostIds } from '../moderation/report';
-import { getAvatarUrlForUser, getDisplayNameForUser, getUsernameForUser } from '../user_identity';
+import { getDisplayNameForUser } from '../user_identity';
 import { getAttendeeCount } from './attendance';
+import {
+  getEventRecIdsForUser,
+  recommendationScoresFromRecIds,
+  serializeEvent,
+  serializeEventFeedPost,
+} from './utils';
 
-export function latLngToH3Index(lat: number, lng: number, resolution: number = 9): string {
-  if (lat < -90 || lat > 90) {
-    throw new RangeError(`Latitude must be between -90 and 90. Got: ${lat}`);
-  }
-  if (lng < -180 || lng > 180) {
-    throw new RangeError(`Longitude must be between -180 and 180. Got: ${lng}`);
-  }
-  if (!Number.isInteger(resolution) || resolution < 0 || resolution > 15) {
-    throw new RangeError(`Resolution must be an integer between 0 and 15. Got: ${resolution}`);
-  }
-  return latLngToCell(lat, lng, resolution);
-}
-
-async function serializeEvent(ctx: QueryCtx, event: Doc<'events'>, recommendationScore?: number) {
-  const [attendeeCount, eventTagLinks] = await Promise.all([
-    getAttendeeCount(ctx, event._id),
-    ctx.db
-      .query('eventTags')
-      .withIndex('by_event', (q) => q.eq('eventId', event._id))
-      .collect(),
-  ]);
-
-  const tags = await Promise.all(eventTagLinks.map(async (link) => await ctx.db.get(link.tagId)));
-
-  return {
-    id: event._id,
-    name: event.name,
-    caption: event.caption,
-    tags: tags.flatMap((tag) => (tag ? [tag.name] : [])),
-    location: event.location,
-    attendeeCount,
-    startDate: event.startDate,
-    endDate: event.endDate,
-    mediaId: event.mediaId ?? null,
-    hostIds: event.hostIds,
-    recommendationScore,
-  };
-}
+export { latLngToH3Index } from './utils';
 
 async function serializeExternalEvent(
   ctx: QueryCtx,
@@ -73,55 +39,19 @@ async function serializeExternalEvent(
   };
 }
 
-async function serializeEventFeedPost(
-  ctx: QueryCtx,
-  post: Doc<'posts'>,
-  viewerId?: Doc<'users'>['_id']
-) {
-  const mediaIds = post.mediaIds ?? [];
-
-  const [author, comments, likes, event] = await Promise.all([
-    ctx.db.get(post.authorId),
-    getThreadedCommentsByPost(ctx, post._id),
-    ctx.db
-      .query('likes')
-      .withIndex('by_postId', (q) => q.eq('postId', post._id))
-      .collect(),
-    post.eventId ? ctx.db.get(post.eventId) : Promise.resolve(null),
-  ]);
-
-  return {
-    id: post._id,
-    caption: post.caption ?? '',
-    creationTime: post._creationTime,
-    authorId: post.authorId,
-    authorName: getDisplayNameForUser(author),
-    authorUsername: getUsernameForUser(author),
-    authorAvatarUrl: getAvatarUrlForUser(author),
-    likes: post.likeCount ?? likes.length,
-    liked: viewerId ? likes.some((like) => like.userId === viewerId) : false,
-    mediaIds,
-    eventId: post.eventId ?? null,
-    eventName: event?.name ?? '',
-    commentCount: countComments(comments),
-    comments,
-  };
-}
-
-function countComments(comments: Awaited<ReturnType<typeof getThreadedCommentsByPost>>): number {
-  return comments.reduce((total, comment) => total + 1 + countComments(comment.replies), 0);
-}
-
 export const getEvents = query({
   args: {},
   handler: async (ctx) => {
-    const [, guestMode] = await __backend_only_guestOrAuthenticatedUser(ctx);
-
+    const [viewer, guestMode] = await __backend_only_guestOrAuthenticatedUser(ctx);
     const events = await ctx.db.query('events').withIndex('by_startDate').collect();
+
+    const shouldLoadRecommendations = !guestMode && viewer !== null;
+    const scoreByEventId = shouldLoadRecommendations
+      ? recommendationScoresFromRecIds(await getEventRecIdsForUser(ctx, viewer._id))
+      : undefined;
+
     return await Promise.all(
-      events.map((event, index) =>
-        serializeEvent(ctx, event, guestMode ? undefined : 1 / (index + 1))
-      )
+      events.map((event) => serializeEvent(ctx, event, scoreByEventId?.get(event._id)))
     );
   },
 });
